pyproject.toml
[project]
name = "finance-news-sentiment-classifier"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.37.8",
    "python-dotenv>=1.0.1",
    "sagemaker>=2.241.0",
]
